Amend to order queries
I used some of these in a document that created a xml file. I kept running into syntax errors and found that in `isCompleted` and `isPaid` there needed to be a param of true set for it to work. Once I did that the feed showed and completed and paid orders were shown